Compute the maxYLabelWith using the precision.
As I introduced the precision, the maxYLabelWith is not the width of the raw
value but the width of the formatted one.

As the computing the of maxYLabelWidth is done in calc(), I moved yInterval in
it because I need it to compute the precision. Also I think it makes more sense,
because the calc() method is supposed to "do any size-related calculations".