[BAR CHART] Fix axis labels when rotation
On a Bar Chart, I needed to show the labels with a 45 degrees angle. However the SVG zone is not hight enough to show the full label (there are cut off). So I did some changes to make sure the zone is big enough to show the text. See the result there : http://img15.hostingpics.net/pics/722911beforeafter.png